CompareMetricBySplit {OurTools}	R Documentation

CompareMetricBySplit

Description

compares one metric variable by a split variable with t.test-oneway.test and wilcox.test-kruskal.test. For two groups confidence intervals for mean difference, difference in pseudomedians (Hodge-Lehmann), and proversion probabilities are provided.

Usage

CompareMetricBySplit(DFR, METRIC, SPLIT, DIGITS = 3, CONF.LEVEL = 0.95)

Arguments

DFR,
is a dataframe

METRIC,
name of metric variable in DFR

SPLIT,
name of split factor (grouping variable) in DFR

DIGITS
= 3, Number of digits for estimates

CONF.LEVEL
= 0.95, Confidence Level for confidence interval

Details

CompareMetricBySplit

Value

a data.frame with results

Author(s)

Dirk Hasenclever 2017-12-10

Examples

x<-c(rnorm(100,2,1),rnorm(200,2.4,1))
x[sample(1:300,13)]<-NA
x<-c(rnorm(100,2,1),rnorm(200,2.5,1))
gr2<-factor(c(rep("Gr1",100),rep("Gr2",200)))
gr3<-factor(c(rep("Gr1",100),rep("Gr2",100),rep("Gr3",100)))
dfr<-data.frame(X=x,GR2=gr2,GR3=gr3)# 
CompareMetricBySplit(dfr,METRIC="X",SPLIT="GR2")
CompareMetricBySplit(dfr,METRIC="X",SPLIT="GR3") 

[Package OurTools version 0.27 Index]