DescribeMetricBySplit-deprecated {OurTools}	R Documentation

DescribeMetricBySplit produces a data.frame describing one or many metric variables

Description

This OUT_DATED function produces a data.frame describing one or many metric variables selected from a data.frame DFR by their names in a character vector COLUMNS SPLIT is a group variable which may be left out

Usage

DescribeMetricBySplit(
  DFR,
  COLUMNS = NULL,
  SPLIT = NULL,
  LONGNAMES = NULL,
  SEL = "All",
  DIGITS = 3,
  ALL = T,
  FORTRANSPOSITION = F
)

Arguments

DFR
is a dataframe (if vector, COLUMNS must remain NULL)

COLUMNS
=NULL colnames in DFR to be described (at least one, unless DFR is a vector)

SPLIT
=NULL split factor (variable), if NULL simple description, NULL is default

LONGNAMES
=NULL optional long names for variables specified by COLUMNS

SEL
="ALL" if specified select statistics from c("MEAN", "SD", "P25", "MEDIAN", "P75", "MIN", "MAX", "MAD", "NVALID", "NMISSING")

DIGITS
Number of digits for percentages , digits=1 is default

ALL
=T If TRUE, also overall statistics are calculated

FORTRANSPOSITION
=F Set to T if you later want to transpose the resulting data.frame - VARIABLE stays redundant

Value

a data.frame with statistics for each COLUMN by the group variable SPLIT

See Also

ourPkg-deprecated

[Package OurTools version 0.27 Index]