Df2Csv {OurTools}	R Documentation

Df2Csv writes a data.frame to a csv file

Description

Df2Csv writes a data.frame to a csv file such that it can be opended in excel and excel interprets numerics with "." as decimal

Usage

Df2Csv(
  df,
  FILENAME = NULL,
  TABLECAPTURE = NULL,
  DIGITS = 4,
  WIDTH = 12,
  JUSTIFY = "right",
  ROW.NAMES = T
)

Arguments

df
as dataframe

FILENAME
as the saved name

TABLECAPTURE
as header for table

DIGITS,
4 is default

WIDTH,
width=12 is default

JUSTIFY
, right is default

ROW.NAMES,
TRUE is default

Value

csv file

Author(s)

Dirk Hasenclever

Examples

## Not run: 
test<-data.frame(a=rnorm(10),b=rep(1:2,5), c=rep(rnorm(10)))
Df2Csv(df=test,TABLECAPTURE="Test of data frame output",DIGITS=2)

## End(Not run)

[Package OurTools version 0.27 Index]