FactorsBySplit {OurTools}	R Documentation

produces a data.frame describing one or many factor variables by a grouping factor SPLIT

Description

The function produces a data.frame describing one or many Factor variables selected from a data.frame DFR by their names in a character vector COLUMNS SPLIT is a group variable which may be left out. FactorsBySplit is a wrapper function calling CrossTabs for each COLUMN, rbinding the results.

Usage

FactorsBySplit(
  DFR,
  COLUMNS,
  SPLIT = NULL,
  LONGNAMES = NULL,
  PVALUE = F,
  NOZEROS = T,
  ALL = NULL,
  DIGITS = 1
)

Arguments

DFR
is a dataframe

COLUMNS
colnames in DFR to be described (at least one)

SPLIT
split factor (variable) of same length as DFR, or alternatively column name in DFR as string. If SPLIT=NULL simple factor description, NULL is default

LONGNAMES
=NULL alternative names for the COLUMNS

PVALUE
p-value from Chi2-test FALSE is default

NOZEROS
eliminates factor levels that were not obserevd, TRUE is default

ALL
= NULL If ALL=NULL 'All' columns on the right, if ALL=T on the left and if ALL=F omitted.

DIGITS
Number of digits for percentages , digits=1 is default

Details

FactorsBySplit

Value

a frequency table by group variable

Author(s)

Dirk Hasenclever 2017-03-14, 2018-01-05

See Also

CrossTabs

Examples

## Not run: 
a<-c(rep("M",240),rep("F",170))
b<-c(rep("S1",110),rep("S2",190),rep("S3",110))
c<-factor(b,levels=c("S0","S1","S2","S3"))
gr<-sample(LETTERS[1:2],410,TRUE)
df<-data.frame(a,b,c,gr)

FactorsBySplit(df, COLUMNS = c("a","b","gr"))
FactorsBySplit(df,c("a","b","c"), SPLIT = "gr")
FactorsBySplit(df,c("a","b","c"), SPLIT = "gr",NOZERO=F)
FactorsBySplit(df,c("a","b"), SPLIT = "gr",ALL=NULL,PVALUE=T)
FactorsBySplit(df,c("a","b"), SPLIT = "gr",ALL=F,PVALUE=T)
FactorsBySplit(df,c("a","b"), SPLIT = "gr",ALL=T,PVALUE=T)
FactorsBySplit(df,c("a","b"), SPLIT = "gr",ALL=T,PVALUE=T,DIGITS=3)
FactorsBySplit(df,c("a","b"), SPLIT = "gr",ALL=T,PVALUE=T,
                         LONGNAMES=c("Gender","STAGE"))

## End(Not run)

[Package OurTools version 0.27 Index]