Glm2Df {OurTools}	R Documentation

Glm2Df

Description

Glm2Df produces a table with results of glm models as printable data.frame.

Usage

Glm2Df(MOD, DIGITS = 3)

Arguments

MOD,
fit of a glm or lm regression Model (without random effects)

DIGITS
=3 round to DIGITS

Value

data.frame with Covariates, coef, se(coef), exp(coef), lower.95 upper.95 p.value

Author(s)

Dirk Hasenclever 2017-12-20, 2019-02-05

Examples

## Not run: 
mod<-lm(WBC~HB+SEX+AGE,data=HLProgFactors)
Glm2Df(mod,DIGITS=3)
mod<-glm(SEX~HB+WBC+AGE,data=HLProgFactors,family = "binomial")
Glm2Df(mod,DIGITS=3)

## End(Not run)

[Package OurTools version 0.27 Index]