HLProgFactors {OurTools}	R Documentation

Data form the International Prognostic Factor Project in advanced Hodgkin Lymphoma.

Description

A dataset with patient characteristics and time to Event data form N= 5141 HL patients.

Usage

HLProgFactors

Format

A data frame with 5141 rows and 19 variables:

PATNO

Patient Number

STUDY

Short name of study group providing data

SEX

Sex of patient

AGE

Age at start of treatment

HISTO

Histological subtype

LAP

Laparotomy: CS = no; PS = yes

STAGE

Stage I, II, III, IV

SSYM

Systemic Symptoms: A=no B=yes

OSVtime

Overall Survival Time [months]

OSVstatus

Vital status at SV: 0=alive; 1=dead

FFTFtime

Freedom from treatment failure time [months]

FFTFstatus

Failure of therapy: 0=no; 1=yes

HB

Haemoglobin [g/dl]

HCT

Haematocrit %

ESR

Erythrocyte sedimentation rate mm/h

WBC

WBC [G/l]

LYMPHO

Lymphocyte Count [G/l]

LYMPHPR

Lymphocyte Count in % WBC

ALBUMIN

Albumin

Source

https://www.nejm.org/doi/10.1056/NEJM199811193392104

[Package OurTools version 0.27 Index]