Level2NA {OurTools}	R Documentation

Level2NA

Description

Level2NA recodes one selected LEVEL of a factor X to NA and eliminates it from levels(X). Level2NA can be applied to a factor or to all factors in a data.frame. NOTE: Level2NA has no effect on character vectors.

Usage

Level2NA(X, LEVEL, EXCEPT = NULL)

Arguments

X
is either a factor or a data.frame

LEVEL
which is to be recoded to NA and has to be specified

EXCEPT
=NULL list colnames of a df that should not be transformed

Value

the transformed factor or data.frame

Author(s)

Dirk Hasenclever 2017-03-22, 2018-01-09

Examples

## Not run: 
# Applied to a vector
Y<-factor(c("BALI","BALI","SUMATRA","LOMBOK","n.n.", "JAVA"))
Y
Level2NA(Y, LEVEL="n.n.")

# Applied to a data.frame
DFR<-data.frame(X=1:6,
                Y=c("BALI","BALI","SUMATRA","LOMBOK","n.n.", "JAVA"),
                Z=factor(c(LETTERS[1:5],"n.n.")))
DFR
DFR<-Level2NA(DFR, LEVEL="BALI")
DFR
DFR<-Level2NA(DFR, LEVEL="n.n.",EXCEPT="Z")
DFR

## End(Not run)

[Package OurTools version 0.27 Index]