LocationScaleEstimates {OurTools}	R Documentation

LocationScaleEstimates

Description

LocationScaleEstimates estimates mean, pseudomedian, sd, mad and mode (of density) of a numeric variable. Estimates and CIs are based on t.test and wilcox.test for mean and pseudomedian. For other parameters the CI is based on bca-bootstrap relying on package boot. CAUTION: Bootstrap may take a while with large data.

Usage

LocationScaleEstimates(
  METRIC,
  PARAM = c("Mean", "Pseudomedian", "Sd"),
  CONF.LEVEL = 0.95,
  DIGITS = 3,
  MODEADJ = 1,
  NBOOT = 999
)

Arguments

METRIC
is a numeric variable

PARAM
=c("Mean","Pseudomedian","Sd") vector of parameter names. Choose from PARAM=c("Mean","Pseudomedian","Sd","Mad","Mode")

CONF.LEVEL
= 0.95 Confidence Level for confidence interval

DIGITS
=3 Number of digits for estimates

MODEADJ
adjust, default=1 for density for mode estimation

NBOOT
=999 number of bootstrap replicates

Value

data.frame with parameters as rows and estimate, lower and upper 95

Author(s)

Dirk Hasenclever 2017-12-14

Examples

METRIC<-c(rnorm(200),rnorm(100,2,1),10:17)
## Not run: 
LocationScaleEstimates(METRIC)
LocationScaleEstimates(METRIC,PARAM=c("Mean","Pseudomedian","Mode"),MODEADJ = .7)

## End(Not run)

[Package OurTools version 0.27 Index]