MultiParameterTimeCourses {OurTools}	R Documentation

MultiParameterTimeCourses

Description

MultiParameterTimeCourses plots courses of multiple metric parameters over a common x-axis, with multiple y-axis.

Usage

MultiParameterTimeCourses(
  DFR,
  TIME,
  COURSES,
  LEFTRIGHT = F,
  SCALES = NULL,
  XLIM = NULL,
  YLIMMIN = NULL,
  YLIMMAX = NULL,
  HLINES = NULL,
  MAIN = "",
  XLAB = "Time Axis",
  YLABS = NULL,
  LWD = 2,
  LTY = NULL,
  COLS = 1,
  PCH = NULL,
  ...
)

Arguments

DFR
is a dataframe

TIME
colname in DFR of available time points to be plotted on the x-axis

COURSES
vector of colnames in DFR for metric parameters to be plotted

LEFTRIGHT
= F default is all y-axes on the left, if set to TRUE y-axis are placed altervatingly left and right.

SCALES
= NULL, specify type of scale for each course out of c("ID","LOG10","POWER_Z","LOGIT")

XLIM
= NULL, specify limits of the time axis

YLIMMIN
= NULL, specify lower limits of axis of the the metric parameters, a vector of length COURSES

YLIMMAX
= NULL, specify upper limits of axis of the the metric parameters, a vector of length COURSES

HLINES
= NULL, plot horizontal lines in the respective scales, a vector of length COURSES with NA to leave out.

MAIN
= "" Plot Title

XLAB
= "Time Axis" x-axis label

YLABS
= NULL optional long y-labels for each parameter in COURSES.

LWD
= 2 line width, one of all or a vector of length COURSES

LTY
= NULL line type, defaults to 1:K, else one of all or a vector of length COURSES

COLS
= 1 colours for parameters, one of all or a vector of length COURSES

PCH
= NULL Point type for measured values, one of all or a vector of length COURSES

...
specify Z for SCALE = "POWER_Z" and/or "PERCENT" for SCALE = "LOGIT"

Value

lower and upper limit of the last plotted y-axis, useful to place points afterwarts.

Author(s)

Dirk Hasenclever 2020-02-26 Thanks to #https://www.r-bloggers.com/multiple-y-axis-in-a-r-plot/

See Also

PlotAxis

Examples

## Not run: 
TT<-1:10
X1<-1:10 + rnorm(10)
X2<- 10^(1:10 + rnorm(10))
X3<- (1:10 + rnorm(10))^3
X4<- T_Logit((1:10 + rnorm(10))/3,INV=T)
X5<- T_Logit((1:10 + rnorm(10))/3,INV=T)*100

X2[4]<-NA

DFR<-data.frame(TT,X1,X2,X3,X4,X5)

MultiParameterTimeCourses(DFR, TIME="TT", COURSES=c("X1","X2","X3","X4","X5"), LEFTRIGHT = F, 
                          YLIMMIN=NULL, YLIMMAX=NULL,
                          SCALES = NULL,
                          HLINES = NULL,
                          YLABS = NULL, LWD = 2, LTY = NULL, COLS = 1:5, PCH = NULL, 
                          XLAB = "Time Axis",
                          XLIM = NULL,MAIN = "")

Lim<-MultiParameterTimeCourses(DFR, TIME="TT", COURSES=c("X1","X2","X3","X4","X5"), LEFTRIGHT = T, 
               YLIMMIN=c(0,.1,0,.3,40), YLIMMAX=c(12,1e12,2000,1,100),
               SCALES = c("ID","LOG10","POWER_Z","LOGIT","LOGIT"),
               HLINES = c(5,1000,NA,.75,80),
               YLABS = c("Linear","log10","to the power of 1/3","logit","logit in percent"), 
               LWD = 3, LTY = 1, COLS = 1:5, PCH = 15, 
               XLAB = "Time Axis [months]",
               XLIM = NULL, MAIN = "Demo MultiParameterTimeCourses",
               Z=1/3,PERCENT=c(F,F,F,F,T))
# The y-coordinates of theplots remain defined by the last axis plotted.
# This scale is invisably returned. It can be used to place points in the plot afterwarts.
graphics::points(5,Lim[1],bg=2,cex=1.7,pch=21)

## End(Not run)

[Package OurTools version 0.27 Index]