NoUmlaute {OurTools}	R Documentation

NoUmlaute eliminates all German Umlaute and sz

Description

NoUmlaute eliminates all German language specific characters (Umlaute and sz replaced ASCII character combinations) in character and factor vectors preserving the class. The ordering of factor levels is not changed. Applying NoUmlaute to other classes triggers a warning and returns the argument identically. NoUmlaute can be applied to a data.frame! ATTENTION: Do not use apply(DFR,2, NoUmlaute) as it changes all classes to factor.

Usage

NoUmlaute(X)

Arguments

X
character or factor variable or data.frame

Details

NoUmlaute

Value

X with all German language specific characters replaced by ASCI if of class character or factor

Author(s)

Dirk Hasenclever 2019-02-15, 2019-10-28

[Package OurTools version 0.27 Index]