OurTools-deprecated {OurTools}	R Documentation

Deprecated functions in package OurTools.

Description

The functions listed below are deprecated and should no longer be used. When possible, alternative functions with similar functionality are mentioned. These functions still work in order to protect old R-scripts, but will not be developed further. Help pages for deprecated functions are available at help("<function>-deprecated").

Usage

DescribeMetricBySplit(
  DFR,
  COLUMNS = NULL,
  SPLIT = NULL,
  LONGNAMES = NULL,
  SEL = "All",
  DIGITS = 3,
  ALL = T,
  FORTRANSPOSITION = F
)

MetricPlotBySplit(
  METRIC,
  SPLIT = NULL,
  TYPE = "Ecdf",
  XLAB = NULL,
  XLIM = NULL,
  MAIN = NULL,
  COL = NULL,
  ADJ = 1,
  METHOD = "jitter"
)

findmode(x, ADJ = 1)

nice.hist(
  x,
  HISTCOL = 7,
  DENSCOL = 4,
  XLAB = NULL,
  YLAB = "Density",
  NCLASS = 40,
  ADJ = 1,
  MAIN = "Histogram with Density",
  XLIM = NULL,
  XAXP = NULL,
  DIGITS = 3,
  LEGEND = F,
  RUG = T,
  LOG10 = F,
  MADFILTER = 0,
  ANALYSISSTRING = "",
  CUTLINE = NULL
)

DescribeMetricBySplit

For DescribeMetricBySplit, use MetricsBySplit.

MetricPlotBySplit

For MetricPlotBySplit, use PlotMetricBySplit.

findmode

For findmode, use ModalValue.

nice.hist

For nice.hist, use NiceHist.

Author(s)

Dirk Hasenclever 2017, 2018-01-31, 2018-09-13, 2019-02-06

Dirk Hasenclever 2017-03-09, 2019-02-06

Dirk Hasenclever 2019-02-06

Dirk Hasenclever 2017-01-22, 2017-03-22, 2019-02-06

See Also

MetricsBySplit

hist

Examples

## Not run: 
x<-c(rnorm(100,2,1),rnorm(200,3,1))
y<-c(rnorm(100,5,1),rnorm(200,7,1))
x[sample(1:300,13)]<-NA
gr<-factor(c(rep("Gr1",100),rep("Gr2",200)))
s<-rep(c(1,2,3),100)
s1<-factor(rep(1,300))
dfr<-data.frame(X=x,Y=y,GR=gr,S=s,S1=s1)

DescribeMetricBySplit(dfr,COLUMNS=c("X","Y"),
              LONGNAMES=c("Variable1","Variable2"),dfr$S1)
DescribeMetricBySplit(dfr,COLUMNS=c("X","Y"),
              LONGNAMES=c("Variable1","Variable2"),DIGITS=1)
DescribeMetricBySplit(dfr,COLUMNS=c("X","Y"),dfr[,"S"])

DescribeMetricBySplit(dfr,COLUMNS=c("X","Y"),dfr$GR,
              SEL=c("MEDIAN","SD","MEAN","NVALID"))

DescribeMetricBySplit(dfr$X,SPLIT=gr,LONGNAMES="Variable1")
DescribeMetricBySplit(dfr$X,COLUMNS=c("X"),gr,LONGNAMES="Variable1")
DescribeMetricBySplit(dfr,COLUMNS=c("X","Y"),SPLIT=gr,
                      LONGNAMES=c("Variable1","Variable2"),SEL="MEDIAN")

## End(Not run)
## Not run: 
METRIC<-c(rnorm(100,10,2),rnorm(200,12,4))
SPLIT<-c(rep("Good",100),rep("Bad",100),rep("DEVELISH",100))

MetricPlotBySplit(METRIC,SPLIT,XLAB="test variable",COL=c(2,3,4),TYPE="Ecdf")
MetricPlotBySplit(METRIC,SPLIT,XLAB="test variable",COL=c(2,3,4),TYPE="Dens",
                  ADJ=c(1,.5,2))
MetricPlotBySplit(METRIC,SPLIT,XLAB="test variable",COL=c(2,3,4),TYPE="Boxplot")
MetricPlotBySplit(METRIC,SPLIT,XLAB="test variable",COL=c(2,3,4),
                  TYPE="Stripchart",METHOD="jitter")

## End(Not run)
## Not run: 
Metric<-c(rnorm(200),rnorm(100,3,1))
nice.hist(Metric)
nice.hist(Metric,RUG=F)
nice.hist(Metric,LEGEND = T)
nice.hist(Metric,LEGEND = T,LOG10=T)
# Dealing with outliers and missing values
Metric<-c(rnorm(200),rnorm(100,3,1),runif(10,5,100),rep(NA,5))
nice.hist(Metric)
SEL<-nice.hist(Metric,MADFILTER=5)
Metric[SEL]
SEL<-nice.hist(Metric,MADFILTER=5,LEGEND=T,LOG10=F,ADJ=.7)

## End(Not run)

[Package OurTools version 0.27 Index]