PlotAxis {OurTools}	R Documentation

PlotAxis

Description

Function to plot a transformed axis. Available TYPEs are ID, POWER_Z, LOG10, LOGIT.

Usage

PlotAxis(
  x,
  AXIS = 1,
  SCALE = "ID",
  LIM = NULL,
  NPRETTY = 8,
  UNIT = NULL,
  LINE = NA,
  COL = 1,
  ...
)

Arguments

x
is the untransformed metric variable to be plotted

AXIS
=1 Choose which axis 1=x-axis, 2=y-axis

SCALE
="ID" chooses between ID, POWER_Z, LOG10, LOGIT

LIM
=NULL plot interval for axis i.e. xlim or ylim

NPRETTY
= 8 influences approximate number of labels when using ID or POWER_Z (cf. base pretty)

UNIT
=NULL if SCALE ="ID" AND diff(LIM) divisable by UNIT plot interval will have UNIT based ticks

LINE
= NA will be passed to axis as line=LINE

COL
= 1 will be passed to axis as col = COL to colour the axis

...
further parameters for scale transformations can be provided via ... Z specifies the power to use with POWER_Z when using POWER_Z PERCENT =F Use percent instead of proportions in LOGIT when using LOGIT

Value

plots the chosen axis in the currently open plot.

Author(s)

Dirk Hasenclever 2017-03-29, 2018-09-12, 2020-02-15 (LINE parameter)

Examples

## Not run: 
x<-rnorm(300,10,3)
hist(x,xaxt="n")
PlotAxis(x,AXIS=1,SCALE="ID",NPRETTY=12)
hist(x,xaxt="n")
PlotAxis(x,AXIS=1,SCALE="ID",NPRETTY=3)

x<-rnorm(300,10,6)^(3/2)
hist(x^(2/3),xaxt="n")
PlotAxis(x,AXIS=1,SCALE="POWER_Z", Z=2/3)

x<-exp(rnorm(700,1,1))
hist(log10(x),xaxt="n")
PlotAxis(x,AXIS=1,SCALE="LOG",LIM=NULL,NPRETTY=2)

x<-rbeta(300,12,1)
hist(T_Logit(x),xaxt="n",nclass=40)
PlotAxis(x,AXIS=1,SCALE="LOGIT",NPRETTY=8)
x<-rbeta(300,3,1)*100
hist(T_Logit(x,PERCENT=T),xaxt="n",nclass=40)
PlotAxis(x,AXIS=1,SCALE="LOGIT",NPRETTY=8,PERCENT=T)

## End(Not run)

[Package OurTools version 0.27 Index]