PlotBinaryByMetric {OurTools}	R Documentation

Plot Binary versus Metric

Description

Basic plot to look at the dependence of a binary variable from a metric one. Scatterplot BINARY verus METRIC. BINARY is jittered to avoid overplotting. Blue line is logistic regression curve. Grey line is lowess smoother. Legend details Nvalid, Baserate and Wald pValue of METRIC in logistic regession.

Usage

PlotBinaryByMetric(
  BINARY,
  METRIC,
  BNAME = NULL,
  MNAME = NULL,
  MAIN = "",
  SCALE = "ID",
  XLIM = NULL,
  NPRETTY = 6,
  SPAN = 0.75,
  LOWESSCOL = 8,
  ...
)

Arguments

BINARY
is a variable of any type with only two values

METRIC
is a metric variable

BNAME
= NULL is an optional label for BINARY

MNAME
= NULL is an optional label for M

MAIN
= "" is an optional title

SCALE
="ID" chooses a ScaleTransformation ID, POWER_Z, LOG10, LOGIT for X: use optional parameter Z to specify the power to use with POWER_Z. use optional parameter PERCENT = T for percent instead of proportions on LOGIT scale.

XLIM
= NULL sets the x-axis range to be plotted manually - generally not necessary

NPRETTY
= 6 influences approximate number of labels when using ID or POWER_Z (cf. base pretty)

SPAN
= 0.75 is the smoothing parameter for lowess

LOWESSCOL
= 8 Colour of Lowess-Smoother. Set to 0 to eliminate the curve

...
Specify Z or PERCENT for SCALE = "POWER_Z" or "LOGIT"

Value

Generates a plot and returns the summary of the logistic regression.

Author(s)

Dirk Hasenclever old; new version 2019-08-09

Examples

METRIC<-rnorm(200); BINARY<-(METRIC+rnorm(200))>0; PlotBinaryByMetric(BINARY,METRIC)

[Package OurTools version 0.27 Index]