PlotDepMetrics {OurTools}	R Documentation

PlotDepMetrics

Description

Produces plots of several dependent METRIC variables on the same scale. The metric variables are specified as COLUMNS of a data.frame DFR Available TYPEs are Ecdf, Dens, Boxplot, Stripchart as in PlotMetricBySplit.

Usage

PlotDepMetrics(
  DFR,
  COLUMNS,
  TYPE = "Ecdf",
  SCALE = "ID",
  MLAB = NULL,
  SPLITNAMES = NULL,
  MLIM = NULL,
  UNIT = NULL,
  MAIN = NULL,
  NVALID = T,
  COL = NULL,
  DRAW_STEPS = F,
  ADJ = 1,
  METHOD = "jitter",
  NPRETTY = 8,
  ...
)

Arguments

DFR
data frame

COLUMNS
colnames of dependent metric variables to be plotted

TYPE
="Ecdf chooses between "Ecdf", "Dens", "Boxplot", "Stripchart"

SCALE
="ID" chooses a ScaleTransformation ID, POWER_Z, LOG10, LOGIT

MLAB
=NULL optional, otherwise derived from call METRIC

SPLITNAMES
=NULL optional names instead of COLUMNS for metric variables;

MLIM
=NULL optional ON THE ORIGINAL SCALE, otherwise derived from pretty

UNIT
=NULL if SCALE ="ID" AND diff(MLIM) divisable by UNIT plot interval will have UNIT based ticks

MAIN
= NULL plot title via main=

NVALID
=T logical switch whether to show number of valid observations per SPLIT group

COL
=NULL colours for the levels of Split - otherwise 2:(G+1)

DRAW_STEPS
=F Set to TRUE for drawing the ecdf steps in full

ADJ
=1 Density smoothing parameter relative to density standard

METHOD
="jitter" METHOD for stripchart

NPRETTY
=8 influences approximate number of labels when using ID or POWER_Z (cf. base pretty)

...
further parameters for scale transformations can be provided via ... Z specifies the power to use with POWER_Z when using POWER_Z PERCENT =F Use percent instead of proportions in LOGIT when using LOGIT

Value

Generates a plot.

Author(s)

Dirk Hasenclever 2017-03-09; 2017-03-20 with scale options, 2017-04-02, 2018-09-12

Examples

## Not run: 
x<-rnorm(100)
y<-rnorm(100,1,1)
z<-rnorm(100,2,3)
dfr<-data.frame(x,y,z)
PlotDepMetrics(DFR=dfr,COLUMNS=colnames(dfr),TYPE="Ecdf",SCALE="ID",
      MLAB=NULL,MLIM=NULL,
      MAIN=NULL,COL=NULL,NVALID=T,ADJ=1,METHOD="jitter",NPRETTY=8)
PlotDepMetrics(DFR=dfr,COLUMNS=colnames(dfr),TYPE="Ecdf",SCALE="ID",
      MLAB=NULL,MLIM=NULL,DRAW_STEPS =  T, 
      MAIN=NULL,COL=NULL,NVALID=T,ADJ=1,METHOD="jitter",NPRETTY=8)

## End(Not run)

[Package OurTools version 0.27 Index]