ProversionProb {OurTools}	R Documentation

ProversionProb

Description

ProversionProb calculates a confidence interval for the proversion probability pr(X>Y)+0.5*pr(X=Y) using method 5 of Newcombe RG (2006). Confidence intervals for an effect size measure based on the Mann-Whitney statistic. Part 1: General issues and tail-area based methods. Stat Med 25(4):543-557 Newcombe RG (2006); Confidence intervals for an effect size measure based on the Mann-Whitney statistic. Part 2: Asymptotic methods and evaluation. Stat Med 25(4):559-573

Usage

ProversionProb(X, Y, CI_LEVEL = 0.95, DIGITS = 7)

Arguments

X, Y
metric or at least ordinal numeric

CI_LEVEL
=0.95 desired confidence level

DIGITS
=7 number of digits

Value

Vector of estimated ProversionProbability with Lower, Upper of CI_LEVEL confidence interval

Author(s)

Dirk Hasenclever 2010-04-28, 2017-12-12, 2018-08-23

Examples

## Not run: 
X<-rnorm(30,10,1)
Y<-rnorm(30,9.5,1)
PlotMetricBySplit(c(X,Y),c(rep("X",length(X)),rep("Y",length(Y))))
ProversionProb(X,Y,DIGITS=4)

## End(Not run)

[Package OurTools version 0.27 Index]