ReconcileLists {OurTools}	R Documentation

ReconcileLists

Description

used to compare and reconsile two lists.

Usage

ReconcileLists(OldDFR, NewDFR, KEYS, ONLYNEW = F)

Arguments

OldDFR
is an old list possibly amended by data managment

NewDFR
is a newly created, updated list

KEYS
is a vector of variable names both in OLDDFR and NEWDFR which identify the entries.

ONLYNEW
= F logical switch to only filter new entries

Value

returns an updated list either with onlynew entries (ONLYNEW=T)or with new and still unresolved entries.

Author(s)

Annett Schrock & Dirk Hasenclever

[Package OurTools version 0.27 Index]