SlidingWindow {OurTools}	R Documentation

SlidingWindow

Description

SlidingWindow time trends of arbitrary statistic.

Usage

SlidingWindow(
  DFR,
  DATE,
  ARGUMENTS,
  STATISTIC,
  WINDOW = 30,
  BY = 7,
  STARTDATE = NULL,
  ENDDATE = NULL,
  PLOT = T,
  MAIN = "Sliding Window Plot",
  XLAB = NULL,
  YLAB = NULL,
  UNIT = 30,
  YLIM = NULL,
  COL = 4,
  ...
)

Arguments

DFR
a data.frame containing a DATE variable and ARGUMENTS for the STATISTIC

DATE
name of a date variable in DFR of class POSIXct or Date, or character of form "YYYY-MM-DD" or "YYYY/MM/DD".

ARGUMENTS
vector of names of those variables in DFR that are used by STATISTIC function; can be set to NULL.

STATISTIC
user supplied function evaluating a data.frame restricting DFR to ARGUMENTS

WINDOW
= 30 width of the sliding window in days

BY
= 7 shift to the next window in days

STARTDATE
= NULL. If not specified, set to the minimum date encountered, ignoring time part. Use "YYYY-MM-DD" format to specify.

ENDDATE
= NULL. If not specified, set to the maximum date encountered, ignoring time part. Use "YYYY-MM-DD" format to specify.

PLOT
= T logical whether a plot is desired

MAIN
= "Sliding Window Plot" is an optional title, to suppress set MAIN = ""

XLAB
= NULL if left NULL, set to Days or years since earliest DATE

YLAB
= NULL if left NULL, set to the name of the supplied STATISTIC function.

UNIT
= 30 days between ticks on time axis. NOTE: UNIT Will be doubled while more than 12 ticks result. Set UNIT = NULL to switch scale to years of 365 days.

YLIM
= NULL optional control over limits of yaxis

COL
= 4 set colour of sliding curve

...
further arguments passed to STATISTIC, e.g. na.rm = T

Value

Returns a data.frame with statistics at sliding windows, and generates a Plot if PLOT = T, plotting values of the statistic at window mids.

Author(s)

Dirk Hasenclever 2020-06-05, Norbert Koehler 2020-06-08, DH 2020-06-08, DH 2020-06-23

[Package OurTools version 0.27 Index]