SuccessRate {OurTools}	R Documentation

SuccessRate

Description

SuccessRate calculates the Wilson-confidence interval for a sucess-rate (using prop.test).

Usage

SuccessRate(FAC, SUCCESS = 2, CONF.LEVEL = 0.95, DIGITS = 1)

Arguments

FAC
is a factor. Will be converted to a factor if not a factor.

SUCCESS
=2, factor level that defines success. If numeric, the respective level of FAC is assumed.

CONF.LEVEL
=0.95, desired confidence level

DIGITS
=7 number of digits

Value

data.frame with SuccessCategory, Number of Successes, Nvalid, Rate estimate in procent and lower and upper of the CONF.LEVEL confidence interval.

Author(s)

Dirk Hasenclever, 2017-12-12

Examples

## Not run: 
SuccessRate(rbinom(453,1,0.33)) 
x<-sample(c("Male", "Female","Inter"),453,replace=T, prob = c(.503,.494,.003)) 
x<-factor(x,levels = c("Male", "Female","Inter")) 
table(x) 
SuccessRate(x,DIGITS=2,SUCCESS=3)
SuccessRate(x,DIGITS=3,SUCCESS=3,CONF.LEVEL=.9)

## End(Not run)

[Package OurTools version 0.27 Index]