SurvivalAtT {OurTools}	R Documentation

SurvivalAtT

Description

Read off survival rates and corresponding CIs at time T from a survfit object.

Usage

SurvivalAtT(TIMEPOINT, SVOBJECT, DIGITS = 7)

Arguments

TIMEPOINT
is the time point at which to read off the Kaplan Meier Kurves is a factorial variable

SVOBJECT,
Survival Object as produced by survfit()

DIGITS
=4, number of digits in output

Details

SurvivalAtT

Value

dataframe with SVestimate, SE, lower95.CI, upper95.CI for each arm/strata

Author(s)

Dirk Hasenclever, 2018-02-26

Examples

## Not run: 
### Test SurvivalAtT
Times<-c(rexp(100,1),rexp(100,.5))*12
Status<-c(rbinom(100,1,.8),rbinom(100,1,.2))
gr<-sample(c("A","B"),200,replace=T)
mod<-survival::survfit(survival::Surv(Times,Status)~gr)
plot(mod)
SurvivalAtT(40,mod)
mod<-survival::survfit(survival::Surv(Times,Status)~1)
SurvivalAtT(40,mod)

## End(Not run)

[Package OurTools version 0.27 Index]