TrimStr {OurTools}	R Documentation

TrimStr removes leading or trailing blanks within a string

Description

Removes spaces within a string variable with leading or trailing blanks. Blanks inbetween are preserved.

Usage

TrimStr(str)

Arguments

str
is a string variable

Value

string variable without leading or trailing blanks

Author(s)

Dirk Hasenclever 2017-03-22

Examples

## Not run: 
test TrimStr
str<-"   Apollo 13 "
str<-" up to date         "
TrimStr(str)

## End(Not run)

[Package OurTools version 0.27 Index]