findmode-deprecated {OurTools}	R Documentation

identifies the modal value of a numeric variable

Description

The function determines the modal value of a numeric variable, which is needed for example in graphics.

Usage

findmode(x, ADJ = 1)

Arguments

x
is a numeric variable

ADJ
adjust, default=1

Value

mode as modal value of a numeric variable

#'

See Also

ourPkg-deprecated

[Package OurTools version 0.27 Index]